fix device in general and fix tests

+++ src/transformers/modeling_layers.py
@@ -536,9 +536,9 @@ class MtpModel(PreTrainedModel):
             # Append the drafted logits
             drafted_logits.append(logits)
             # Decode one token
-            next_token_logits = logits[:, -1, :].to(device=input_ids.device, dtype=torch.float32)
+            next_token_logits = logits[:, -1, :].to(device=input_ids.device)
             if logits_processor is not None and full_input_ids is not None:
-                next_token_logits = logits_processor(full_input_ids, next_token_logits)
+                next_token_logits = logits_processor(full_input_ids, next_token_logits.to(dtype=torch.float32))
             if do_sample:
                 probs = nn.functional.softmax(next_token_logits, dim=-1, dtype=torch.float32)
                 next_mtp_token = torch.multinomial(probs, num_samples=1)
